Split up config.py in de and ttsim part.
