+++ pyproject.toml
@@ -247,17 +247,15 @@ extend-ignore = [
 exclude = []
 
 [tool.ruff.lint.per-file-ignores]
-"src/_gettsim_tests/test_rounding.py" = ["PT019"]
-"src/_gettsim/functions/all_functions_for_docs.py" = ["F401", "E501", "F403"]
-"src/_gettsim/typing.py" = ["PGH", "PLR", "SIM114"]
-"src/_gettsim/piecewise_functions.py" = ["PLR"]
-"src/_gettsim/policy_environment.py" = ["S506", "PLR"]
-"src/_gettsim/transfers/*" = ["E501"]
-"src/_gettsim/taxes/*" = ["E501"]
-"src/_gettsim/taxes/familie/familie.py" = ["E501"]
-"src/_gettsim/visualization.py" = ["PLR"]
-"src/_gettsim_tests/test_synthetic.py" = ["E501"]
-"src/_gettsim/config.py" = ["E501"]
+"src/_gettsim_tests/ttsim/test_rounding.py" = ["PT019"]
+"src/_gettsim/ttsim/functions/all_functions_for_docs.py" = ["F401", "E501", "F403"]
+"src/_gettsim/ttsim/typing.py" = ["PGH", "PLR", "SIM114"]
+"src/_gettsim/ttsim/piecewise_functions.py" = ["PLR"]
+"src/_gettsim/ttsim/policy_environment.py" = ["S506", "PLR"]
+"src/_gettsim/de/*" = ["E501"]
+"src/_gettsim/ttsim/visualization.py" = ["PLR"]
+"src/_gettsim_tests/de/test_synthetic.py" = ["E501"]
+"src/_gettsim/de/config.py" = ["E501"]
 "docs/**/*.ipynb" = ["T201"]
 
 [tool.ruff.lint.pydocstyle]
