+++ src/_gettsim/de/einkommensteuer/abzüge/altersfreibetrag.py
@@ -1,6 +1,6 @@
 """Tax allowances for the elderly."""
 
-from _gettsim.config import numpy_or_jax as np
+from _gettsim.ttsim.config import numpy_or_jax as np
 from _gettsim.ttsim.function_types import policy_function
