+++ src/_gettsim/de/einkommensteuer/abzüge/pauschbetrag_behinderung.py
@@ -1,6 +1,6 @@
 """Tax allowances for the disabled."""
 
-from _gettsim.config import numpy_or_jax as np
+from _gettsim.ttsim.config import numpy_or_jax as np
 from _gettsim.ttsim.function_types import policy_function
