+++ src/_gettsim/de/wohngeld/einkommen.py
@@ -1,7 +1,7 @@
 """Income relevant for housing benefit calculation."""
 
 from _gettsim.ttsim.aggregation import AggregateByPIDSpec
-from _gettsim.config import numpy_or_jax as np
+from _gettsim.ttsim.config import numpy_or_jax as np
 from _gettsim.ttsim.function_types import policy_function
 from _gettsim.ttsim.piecewise_functions import piecewise_polynomial
