+++ src/_gettsim/ttsim/config.py
@@ -1,0 +1,37 @@
+from __future__ import annotations
+
+import importlib
+from pathlib import Path
+
+import numpy
+
+# Defaults
+USE_JAX = False
+numpy_or_jax = numpy
+
+
+def set_array_backend(backend: str):
+    """Set array library backend.
+
+    backend (str): Must be in {'jax', 'numpy'}.
+
+    """
+    if backend not in {"jax", "numpy"}:
+        raise ValueError(f"Backend must be in {'jax', 'numpy'} but is {backend}.")
+
+    if backend == "jax":
+        assert importlib.util.find_spec("jax") is not None, "JAX is not installed."
+        global USE_JAX  # noqa: PLW0603
+        global numpy_or_jax  # noqa: PLW0603
+        import jax
+
+        USE_JAX = True
+        numpy_or_jax = jax.numpy
+        jax.config.update("jax_platform_name", "cpu")
+
+
+# Obtain the root directory of the package.
+RESOURCE_DIR = Path(__file__).parent.resolve()
+
+# Path to the taxes and transfers functions
+PATH_TO_FUNCTIONS_ROOT = RESOURCE_DIR / "de"
