+++ src/_gettsim_tests/ttsim/test_config.py
@@ -6,7 +6,7 @@ import _gettsim
 
 
 def test_default_backend():
-    from _gettsim.config import numpy_or_jax
+    from _gettsim.ttsim.config import numpy_or_jax
 
     assert numpy_or_jax.__name__ == "numpy"
 
@@ -15,18 +15,18 @@ def test_set_backend():
     is_jax_installed = importlib.util.find_spec("jax") is not None
 
     # expect default backend
-    from _gettsim.config import numpy_or_jax
+    from _gettsim.ttsim.config import numpy_or_jax
 
     assert numpy_or_jax.__name__ == "numpy"
 
     if is_jax_installed:
         # set jax backend
         _gettsim.config.set_array_backend("jax")
-        from _gettsim.config import numpy_or_jax
+        from _gettsim.ttsim.config import numpy_or_jax
 
         assert numpy_or_jax.__name__ == "jax.numpy"
 
-        from _gettsim.config import USE_JAX
+        from _gettsim.ttsim.config import USE_JAX
 
         assert USE_JAX
     else:
